+++ Source/Networking/VRNetPlayer.cs
@@ -1,4 +1,5 @@
-﻿using System.IO;
+﻿using System.Collections.Generic;
+using System.IO;
 using System.Linq;
 using GameNetcodeStuff;
 using LCVR.Assets;
@@ -16,9 +17,6 @@ namespace LCVR.Networking;
 /// </summary>
 public class VRNetPlayer : MonoBehaviour
 {
-    private ChainIKConstraintData originalLeftArmConstraintData;
-    private ChainIKConstraintData originalRightArmConstraintData;
-
     private GameObject playerGhost;
     private Transform usernameBillboard;
     private CanvasGroup usernameAlpha;
@@ -31,6 +29,9 @@ public class VRNetPlayer : MonoBehaviour
     private Transform rightController;
     private Transform leftHandVRTarget;
     private Transform rightHandVRTarget;
+
+    private TwoBoneIKConstraint leftArmVRRig;
+    private TwoBoneIKConstraint rightArmVRRig;
 
     private HandTargetOverride? leftHandTargetOverride;
     private HandTargetOverride? rightHandTargetOverride;
@@ -62,6 +63,8 @@ public class VRNetPlayer : MonoBehaviour
     public FingerCurler RightFingerCurler { get; private set; }
 
     public PlayerData AdditionalData { get; private set; }
+
+    private readonly List<GameObject> cleanupPool = [];
     
     private void Awake()
     {
@@ -109,6 +112,9 @@ public class VRNetPlayer : MonoBehaviour
         RightItemHolder.SetParent(Bones.RightHand, false);
         RightItemHolder.localPosition = new Vector3(-0.002f, 0.036f, -0.042f);
         RightItemHolder.localEulerAngles = new Vector3(356.3837f, 357.6979f, 0.1453f);
+        
+        cleanupPool.Add(LeftItemHolder.gameObject);
+        cleanupPool.Add(RightItemHolder.gameObject);
 
         // Set up finger curlers
         LeftFingerCurler = new FingerCurler(Bones.LeftHand, true);
@@ -141,7 +147,7 @@ public class VRNetPlayer : MonoBehaviour
 
         usernameText.text = $"<noparse>{PlayerController.playerUsername}</noparse>";
 
-        var networkSystem = VRSession.Instance.NetworkSystem;
+        var networkSystem = NetworkSystem.Instance;
         
         prefsChannel = networkSystem.CreateChannel(ChannelType.PlayerPrefs, PlayerController.playerClientId);
         rigChannel = networkSystem.CreateChannel(ChannelType.Rig, PlayerController.playerClientId);
@@ -176,48 +182,70 @@ public class VRNetPlayer : MonoBehaviour
         Bones.LeftArmRigHint.localPosition = new Vector3(-10f, -2f, -1f);
 
         // Disable built-in constraints since they don't support hints (fucks up the elbows)
-        var originalLeftArmConstraint = Bones.LeftArmRig.GetComponent<ChainIKConstraint>();
-        originalLeftArmConstraintData = originalLeftArmConstraint.data;
-        Destroy(originalLeftArmConstraint);
+        Bones.LeftArmRig.GetComponent<ChainIKConstraint>().weight = 0;
 
-        var leftArmConstraint = Bones.LeftArmRig.gameObject.AddComponent<TwoBoneIKConstraint>();
-        leftArmConstraint.data.root = Bones.LeftUpperArm;
-        leftArmConstraint.data.mid = Bones.LeftLowerArm;
-        leftArmConstraint.data.tip = Bones.LeftHand;
-        leftArmConstraint.data.target = Bones.LeftArmRigTarget;
-        leftArmConstraint.data.hint = Bones.LeftArmRigHint;
-        leftArmConstraint.data.hintWeight = 1;
-        leftArmConstraint.data.targetRotationWeight = 1;
-        leftArmConstraint.data.targetPositionWeight = 1;
+        // Create new hints and targets for the VR rig
+        var leftArmRigHint = new GameObject("VR Left Arm Rig Hint")
+        {
+            transform =
+            {
+                parent = Bones.LeftArmRigHint.parent,
+                localPosition = Bones.LeftArmRigHint.localPosition,
+                localRotation = Bones.LeftArmRigHint.localRotation,
+                localScale = Bones.LeftArmRigHint.localScale
+            }
+        }.transform;
+        
+        cleanupPool.Add(leftArmRigHint.gameObject);
+        
+        leftArmVRRig = Bones.LeftArmRig.gameObject.AddComponent<TwoBoneIKConstraint>();
+        leftArmVRRig.data.root = Bones.LeftUpperArm;
+        leftArmVRRig.data.mid = Bones.LeftLowerArm;
+        leftArmVRRig.data.tip = Bones.LeftHand;
+        leftArmVRRig.data.target = leftHandVRTarget;
+        leftArmVRRig.data.hint = leftArmRigHint;
+        leftArmVRRig.data.hintWeight = 1;
+        leftArmVRRig.data.targetRotationWeight = 1;
+        leftArmVRRig.data.targetPositionWeight = 1;
+        leftArmVRRig.weight = 1;
 
         // Setting up the right arm
 
         Bones.RightArmRigHint.localPosition = new Vector3(12.5f, -2f, -1f);
 
         // Disable built-in constraints since they don't support hints (fucks up the elbows)
-        var originalRightArmConstraint = Bones.RightArmRig.GetComponent<ChainIKConstraint>();
-        originalRightArmConstraintData = originalRightArmConstraint.data;
-        Destroy(originalRightArmConstraint);
+        Bones.RightArmRig.GetComponent<ChainIKConstraint>().weight = 0;
+        
+        // Create new hints and targets for the VR rig
+        var rightArmRigHint = new GameObject("VR Right Arm Rig Hint")
+        {
+            transform =
+            {
+                parent = Bones.RightArmRigHint.parent,
+                localPosition = Bones.RightArmRigHint.localPosition,
+                localRotation = Bones.RightArmRigHint.localRotation,
+                localScale = Bones.RightArmRigHint.localScale
+            }
+        }.transform;
+        
+        cleanupPool.Add(rightArmRigHint.gameObject);
 
-        var rightArmConstraint = Bones.RightArmRig.gameObject.AddComponent<TwoBoneIKConstraint>();
-        rightArmConstraint.data.root = Bones.RightUpperArm;
-        rightArmConstraint.data.mid = Bones.RightLowerArm;
-        rightArmConstraint.data.tip = Bones.RightHand;
-        rightArmConstraint.data.target = Bones.RightArmRigTarget;
-        rightArmConstraint.data.hint = Bones.RightArmRigHint;
-        rightArmConstraint.data.hintWeight = 1;
-        rightArmConstraint.data.targetRotationWeight = 1;
-        rightArmConstraint.data.targetPositionWeight = 1;
+        rightArmVRRig = Bones.RightArmRig.gameObject.AddComponent<TwoBoneIKConstraint>();
+        rightArmVRRig.data.root = Bones.RightUpperArm;
+        rightArmVRRig.data.mid = Bones.RightLowerArm;
+        rightArmVRRig.data.tip = Bones.RightHand;
+        rightArmVRRig.data.target = rightHandVRTarget;
+        rightArmVRRig.data.hint = rightArmRigHint;
+        rightArmVRRig.data.hintWeight = 1;
+        rightArmVRRig.data.targetRotationWeight = 1;
+        rightArmVRRig.data.targetPositionWeight = 1;
+        rightArmVRRig.weight = 1;
 
         GetComponentInChildren<RigBuilder>().Build();
     }
 
     private void Update()
     {
-        // Re-apply animator if necessary
-        if (PlayerController.playerBodyAnimator.runtimeAnimatorController != AssetManager.RemoteVrMetarig)
-            PlayerController.playerBodyAnimator.runtimeAnimatorController = AssetManager.RemoteVrMetarig;
-        
         // Apply crouch offset
         crouchOffset = Mathf.Lerp(crouchOffset, crouchState switch
         {
@@ -339,13 +367,26 @@ public class VRNetPlayer : MonoBehaviour
         if (StartOfRound.Instance.localPlayerController.localVisorTargetPoint is not null)
             usernameBillboard.LookAt(StartOfRound.Instance.localPlayerController.localVisorTargetPoint);
     }
+    
+    public void UpdateIKWeights()
+    {
+        // Constants
+        PlayerController.cameraLookRig1.weight = 0.45f;
+        PlayerController.cameraLookRig2.weight = 1;
+        PlayerController.leftArmRigSecondary.weight = 0;
+        PlayerController.rightArmRigSecondary.weight = 0;
+
+        PlayerController.leftArmRig.weight = 0;
+        PlayerController.rightArmRig.weight = 0;
+        leftArmVRRig.weight = 1;
+        rightArmVRRig.weight = 1;
+    }
 
     /// <summary>
     /// Show the spectator ghost
     /// </summary>
     public void ShowSpectatorGhost()
     {
-        // Show player ghost when player dies
         foreach (var renderer in playerGhost.GetComponentsInChildren<MeshRenderer>())
         {
             renderer.enabled = true;
@@ -420,25 +461,25 @@ public class VRNetPlayer : MonoBehaviour
     /// </summary>
     private void OnDestroy()
     {
-        Destroy(playerGhost);
-        Destroy(xrOrigin.gameObject);
-
         Bones.ResetToPrefabPositions();
 
-        Destroy(Bones.LeftArmRig.GetComponent<TwoBoneIKConstraint>());
-        Destroy(Bones.RightArmRig.GetComponent<TwoBoneIKConstraint>());
-
-        var leftArmConstraint = Bones.LeftArmRig.gameObject.AddComponent<ChainIKConstraint>();
-        var rightArmConstraint = Bones.RightArmRig.gameObject.AddComponent<ChainIKConstraint>();
-
-        leftArmConstraint.data = originalLeftArmConstraintData;
-        rightArmConstraint.data = originalRightArmConstraintData;
-
-        GetComponentInChildren<RigBuilder>().Build();
-
-        PlayerController.playerBodyAnimator.runtimeAnimatorController =
-            StartOfRound.Instance.otherClientsAnimatorController;
+        // Make sure to destroy immediately, otherwise the rig rebuilding will still use the constraints and cause errors
+        DestroyImmediate(Bones.LeftArmRig.GetComponent<TwoBoneIKConstraint>());
+        DestroyImmediate(Bones.RightArmRig.GetComponent<TwoBoneIKConstraint>());
         
+        Bones.LeftArmRig.GetComponent<ChainIKConstraint>().weight = 1;
+        Bones.RightArmRig.GetComponent<ChainIKConstraint>().weight = 1;
+
+        // Rebuild rig now that we changed up the constraints
+        GetComponentInChildren<RigBuilder>().Build();
+        
+        Destroy(playerGhost);
+        Destroy(xrOrigin.gameObject);
+        
+        foreach (var el in cleanupPool)
+            Destroy(el);
+        
+        cleanupPool.Clear();
         prefsChannel.Dispose();
         rigChannel.Dispose();
         spectatorRigChannel.Dispose();